// frontend/apps/web-antd/src/types/pop-tail.ts
export interface AuthorityInfo {
  ID: number;
  authorityId: number;
  authorityName: string;
  defaultRouter: string;
  enable?: number;
  parentId: number;
  status?: number;
  children: AuthorityInfo[];
  menus?: MenuInfo[];
  userIds?: number[];
}

export interface UserInfo {
  ID: number;
  uuid: string;
  userName: string;
  nickName: string;
  authorityId: number;
  authority: AuthorityInfo;
  authorities: AuthorityInfo[];
  authorityIds?: number[];
  headerImg: string;
  phone: string;
  email: string;
  enable: number;
  originSetting?: Record<string, unknown>;
}
// ...
export interface PolicyPath {
  path: string;
  method: string;
}

export interface PageResult<T> {
  List: T[];
  Total: number;
  Page: number;
  PageSize: number;
}

export interface MenuMeta {
  title: string;
  icon: string;
  activeName?: string;
  closeTab?: boolean;
  defaultMenu?: boolean;
  keepAlive?: boolean;
  transitionType?: string;
  btns?: Record<string, boolean>;
}

export interface MenuParameterInfo {
  ID?: number;
  sysBaseMenuID?: number;
  type: string;
  key: string;
  value: string;
}

export interface MenuButtonInfo {
  ID?: number;
  sysBaseMenuID?: number;
  name: string;
  desc: string;
}

export interface MenuInfo {
  ID: number;
  parentId: number;
  name: string;
  path: string;
  component: string;
  sort: number;
  hidden: boolean;
  meta: MenuMeta;
  children: MenuInfo[];
  authoritys?: AuthorityInfo[];
  menuBtn?: MenuButtonInfo[];
  parameters?: MenuParameterInfo[];
}
// ...
export interface AuthorityButtonMatrixBatchSelectionItem {
  menuID: number;
  selected: number[];
}